Version 0.01.003.0009 - Code analysis changes, including whitespace.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AarreAutonomousDriveByGyro.java
@@ -104,7 +104,7 @@ public class AarreAutonomousDriveByGyro extends LinearOpMode {
 
         // make sure the gyro is calibrated before continuing
         while (!isStopRequested() && gyro.isCalibrating()) {
-            sleep(50);
+	        sleep((long) 50);
             idle();
         }
